Health UI Implementation
- Basic hp bar and animations

+++ Assets/Scripts/Character/Components/Defaults/CharacterHealth.cs
@@ -7,13 +7,13 @@ public class CharacterHealth : Health
     private Character _Character;
 
     // Events
-    public delegate void OnDamage(float amount);
+    public delegate void OnDamage(int amount);
     public static event OnDamage onDamage;
 
-    public delegate void OnHeal(float amount);
+    public delegate void OnHeal(int amount);
     public static event OnHeal onHeal;
 
-    public delegate void OnMaxHealthIncrease(float amount, bool heal);
+    public delegate void OnMaxHealthIncrease(int amount, bool heal);
     public static event OnMaxHealthIncrease onMaxHealthIncrease;
 
     public delegate void OnDeath();
@@ -31,25 +31,25 @@ public class CharacterHealth : Health
     {
         if(_Character && !_Character.IsHitable) return;
         base.Damage(amount, playHurtAnim);
-        onDamage?.Invoke(amount);
+        onDamage?.Invoke((int) amount);
     }
 
     public override void IncreaseMaxHealth(float amount)
     {
         base.IncreaseMaxHealth(amount);
-        onMaxHealthIncrease?.Invoke(amount, false);
+        onMaxHealthIncrease?.Invoke((int) amount, false);
     }
 
     public override void IncreaseMaxHealthAndHeal(float amount)
     {
         base.IncreaseMaxHealth(amount);
-        onMaxHealthIncrease?.Invoke(amount, true);
+        onMaxHealthIncrease?.Invoke((int) amount, true);
     }
 
     public override void Heal(float amount)
     {
         base.Heal(amount);
-        onHeal?.Invoke(amount);
+        onHeal?.Invoke((int) amount);
     }
 
     public override void Die()
